Moved Damage tests to own file
